requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
black==23.3.0
Brotli==1.0.9
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.0.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
gunicorn==20.1.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.13.1
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
platformdirs==3.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.1
PyJWT==2.6.0
pytest==7.2.1
pytest-testdox==3.0.1
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
wcwidth==0.2.6
whitenoise==6.4.0

